backend/requirements.txt
# FastAPI核心
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库
sqlalchemy==2.0.45
psycopg2-binary==2.9.9
alembic==1.12.1
asyncpg==0.31.0

# 认证和安全性
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 环境配置
python-dotenv==1.0.0
pydantic-settings==2.1.0

# 开发工具（macOS特化）
watchfiles==0.21.0      # 文件监控，实现热重载
rich==13.7.0           # 漂亮的终端输出
ipython==8.17.2        # 交互式Python（macOS开发友好）
debugpy==1.8.0         # VS Code调试支持

# 数据验证
pydantic==2.5.0
email-validator==2.1.0

# 实用工具
python-dateutil==2.8.2
python-slugify==8.0.1

# 测试（开发环境）
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2


#日志记录
psutil==5.9.6

#多次测试没有安装成功的
# Failed to build psycopg2-binary asyncpg pydantic-core